removed LoaderContainer
