--- src/components/molecules/loader-container/loader-container.tsx
@@ -1,12 +1,0 @@
-import Loader from '../../atoms/loader/loader.tsx';
-import './loader-container.css';
-
-function LoaderContainer() {
-  return (
-    <div className="loader-container">
-      <Loader />
-    </div>
-  );
-}
-
-export default LoaderContainer;
